Separate extension code
